Factor out ConfigOperation class from Dispatcher (#437)
